Big push, but everything is working without bugs
Made ObjectViewModel a singleton to be used across activities.  There is error checking when switching between the main activity and the detailed monster activity to see if the api returned anything or not.  if it didn't a custom toast message pops up on the main activity.

TODO:: learn to parcelize lixt better.  the data classes are set up as parcelize, however when indexing parcelize list to display information on an activity the error "java.util.arraylist cannot be cast to (object)"

TODO: implement caching